Refactor TombVanillaCompounder into base and child class
New child class is called TombSoloCrypt and is essentially what TombVanillaCompounder used to be. This is in preparation for keeping only general purpose functions inside TombVanillaCompounder so that it can be used as a base class for multiple crypts: one for solo user/operator, another for multiple users and single operator and so on.